+++ tests/src/Unit/HomepageCookieLanguageRedirectionTest.php
@@ -34,6 +34,15 @@ class HomepageCookieLanguageRedirectionTest extends UnitTestCase {
    */
   protected $cookieLanguageRedirectionEventSubscriber;
 
+  /**
+   * Mimic Symfony MASTER_REQUEST/MAIN_REQUEST from HttpKernelInterface.
+   *
+   * In order to keep Drupal 9 compatibility, we need to define this constant by
+   * ourselves as Drupal 11 (Symfony 7) drop usage of MASTER_REQUEST.
+   *
+   * @see https://www.drupal.org/node/3236639
+   */
+  private const  MAIN_REQUEST = 1;
 
   /**
    * The event subscriber to be tested.
@@ -119,7 +128,7 @@ class HomepageCookieLanguageRedirectionTest extends UnitTestCase {
     $kernel = $this->prophesize(HttpKernelInterface::class);
     $request = Request::create('/', 'GET');
 
-    $event = new RequestEvent($kernel->reveal(), $request, HttpKernelInterface::MASTER_REQUEST);
+    $event = new RequestEvent($kernel->reveal(), $request, self::MAIN_REQUEST);
 
     $this->pathMatcher
       ->expects($this->once())
@@ -143,7 +152,7 @@ class HomepageCookieLanguageRedirectionTest extends UnitTestCase {
     $this->request->server->set('HTTP_ACCEPT_LANGUAGE', 'fr');
     $this->request->server->set('HTTP_REFERER', 'https://www.google.ch');
 
-    $event = new RequestEvent($kernel->reveal(), $request, HttpKernelInterface::MASTER_REQUEST);
+    $event = new RequestEvent($kernel->reveal(), $request, self::MAIN_REQUEST);
 
     $currentLanguage = $this->createMock(LanguageInterface::class);
     $currentLanguage->expects($this->never())
@@ -183,7 +192,7 @@ class HomepageCookieLanguageRedirectionTest extends UnitTestCase {
 
     $this->request->server->set('HTTP_ACCEPT_LANGUAGE', 'fr');
 
-    $event = new RequestEvent($kernel->reveal(), $request, HttpKernelInterface::MASTER_REQUEST);
+    $event = new RequestEvent($kernel->reveal(), $request, self::MAIN_REQUEST);
 
     $currentLanguage = $this->createMock(LanguageInterface::class);
     $currentLanguage->expects($this->never())
@@ -224,7 +233,7 @@ class HomepageCookieLanguageRedirectionTest extends UnitTestCase {
     $this->request->server->set('HTTP_ACCEPT_LANGUAGE', 'fr');
     $this->request->cookies->set('home_redirect_lang_preferred_langcode', 'fr');
 
-    $event = new RequestEvent($kernel->reveal(), $request, HttpKernelInterface::MASTER_REQUEST);
+    $event = new RequestEvent($kernel->reveal(), $request, self::MAIN_REQUEST);
 
     $currentLanguage = $this->createMock(LanguageInterface::class);
     $currentLanguage->expects($this->once())
@@ -266,7 +275,7 @@ class HomepageCookieLanguageRedirectionTest extends UnitTestCase {
     $this->request->server->set('HTTP_ACCEPT_LANGUAGE', 'fr');
     $this->request->cookies->set('home_redirect_lang_preferred_langcode', 'ru');
 
-    $event = new RequestEvent($kernel->reveal(), $request, HttpKernelInterface::MASTER_REQUEST);
+    $event = new RequestEvent($kernel->reveal(), $request, self::MAIN_REQUEST);
 
     $currentLanguage = $this->createMock(LanguageInterface::class);
     $currentLanguage->expects($this->once())
@@ -313,7 +322,7 @@ class HomepageCookieLanguageRedirectionTest extends UnitTestCase {
     $this->request->server->set('HTTP_ACCEPT_LANGUAGE', 'fr');
     $this->request->cookies->set('home_redirect_lang_preferred_langcode', 'en');
 
-    $event = new RequestEvent($kernel->reveal(), $request, HttpKernelInterface::MASTER_REQUEST);
+    $event = new RequestEvent($kernel->reveal(), $request, self::MAIN_REQUEST);
 
     $currentLanguage = $this->createMock(LanguageInterface::class);
     $currentLanguage->expects($this->once())